# contracts/integration-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Transit Hub integration API · sandbox reference server",
    "version": "1.0.0",
    "description": "REST + WebSocket /api/v1/stream. Network source is synthetic CSV replay. Dispatch commands execute only in the sandbox. See server/README.md for provider and ML integration."
  },
  "paths": {
    "/api/v1/network/summary": {
      "get": {
        "operationId": "getNetworkSummary",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkSummaryDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/routes": {
      "get": {
        "operationId": "getRoutes",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "risk_levels",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "transport_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutesDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/vehicles": {
      "get": {
        "operationId": "getVehicles",
        "parameters": [
          {
            "name": "route_ids",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "risk_levels",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bbox",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehiclesDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/alerts": {
      "get": {
        "operationId": "getAlerts",
        "parameters": [
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "route_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertsDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/delay-series": {
      "get": {
        "operationId": "getDelaySeries",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bucket",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "route_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DelaySeriesDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/risk-distribution": {
      "get": {
        "operationId": "getRiskDistribution",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskDistributionDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/top-routes": {
      "get": {
        "operationId": "getTopRoutes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopRoutesDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forecast": {
      "get": {
        "operationId": "getForecast",
        "parameters": [
          {
            "name": "horizon_min",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/external/yandex-weather/timeline": {
      "get": {
        "operationId": "getYandexWeatherTimeline",
        "description": "Server-side Yandex GraphQL proxy; 5 minute shared cache. Requires cloudiness/Nowcast entitlement.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "configured",
                    "source",
                    "steps",
                    "fetchedAt"
                  ],
                  "properties": {
                    "configured": {
                      "type": "boolean",
                      "description": "Server key configured; does not guarantee Nowcast entitlement"
                    },
                    "source": {
                      "type": "string",
                      "const": "Яндекс Погода"
                    },
                    "steps": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "time",
                          "generation"
                        ],
                        "properties": {
                          "time": {
                            "type": "integer",
                            "description": "Forecast frame time, Unix seconds"
                          },
                          "generation": {
                            "type": "integer",
                            "description": "Provider generation time, Unix seconds"
                          }
                        }
                      }
                    },
                    "fetchedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "502": {
            "description": "Provider unavailable or entitlement denied"
          },
          "503": {
            "description": "Not configured, disabled or rate limited"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "NetworkSummaryDto": {
        "type": "object",
        "properties": {
          "vehicles_total": {
            "type": "number"
          },
          "vehicles_active": {
            "type": "number"
          },
          "routes_active": {
            "type": "number"
          },
          "on_time_percent": {
            "type": "number"
          },
          "at_risk_percent": {
            "type": "number"
          },
          "delayed_percent": {
            "type": "number"
          },
          "average_delay_sec": {
            "type": "number"
          },
          "average_predicted_delay_sec": {
            "type": "number"
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": [
          "vehicles_total",
          "vehicles_active",
          "routes_active",
          "on_time_percent",
          "at_risk_percent",
          "delayed_percent",
          "average_delay_sec",
          "average_predicted_delay_sec",
          "timestamp"
        ]
      },
...
      "ErrorDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBodyDto"
          }
        },
        "required": [
          "error"
        ]
      },
      "RoutesDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RouteDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
...
      "VehiclesDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VehicleDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "transit_api_session"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    },
    {
      "SessionCookie": []
    },
    {}
  ]
}